Bugfix to Bluejelly runtime.
After some INTENSE debugging, found that bp calculation was wrong for
continuation and catch markers. The bp was being set to sp-1 instead of
sp. That was essentially a copy & paste error from the update marker
case, where bp must be sp-1. I paid the price for implementing updates
first and not using my head when adding continuations and exceptions.

In addition, fixed an error in printList (the last part of the function
was plain wrong), and evalVar, which was pushing stuff to eval before
pushing the continuation marker... fail!

The moral: ALWAYS push continuation first, then build node to eval.
Think things are ok now, but I don't trust my self. I will add an
extensive test suite to the assembler, same way as in L4.

Wow, what a crazy way to start 2013!