Runtime runner as a generic testing utility.
Does not depend on Bluejelly.Runtime, which is ok (we don't want
general purpose utilities to depend on anything specific). Yet,
this class assumes a certain folder organization. I don't think this is
essentially wrong, although I'm not thrilled about putting this class
here. But I haven't a best place so far, so it says here for now.