*Big* refactoring:
Change representation of multiple data structures:

  - Introduce scoped name, they fix a bug in the import chaser (see below)
  - Lift Fixity and ExportInfo types to the common package
  - Use Symbols and ScopedNames instead of Names whenever possible
  - Use a common map for declaratons in a ModDefn

The effect of such a big change is solving a nasty bug in the import chaser:
Suppose we have the module: `module M where data C = C|A', and in another
module we have an import of the form `import A hiding(C(A))'. We end up
having `C' in scope, but since names aren't scoped this `C' refers to both
the tycon and the dcon. With scoped names only a `C' with scope = IdScope
(i.e., only the dcon) ends up in scope.

I will keep refactoring the ModIface, ModDefn, and BjcEnv data structures
so we have a QualName rather than a Name whenever a name *has* to be
qualified. Let's make invalid states irrepresentable.